Added streaming trasnsformer as the subclass of Transformer
To conform to the design pattern, converted Transformer to an abstract base from which the Streaming Transformer is inherited.
